Adding js dynamic imports
