+++ features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/store/.babelrc
@@ -1,7 +1,7 @@
 {
   "presets": [
-    "es2015",
+    "env",
     "react"
   ],
-  "plugins": ["transform-class-properties", "transform-object-rest-spread"]
+  "plugins": ["transform-class-properties", "transform-object-rest-spread", "syntax-dynamic-import"]
 }
