+++ features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/store/.eslintrc.js
@@ -66,5 +66,5 @@ module.exports = {
         'no-else-return': 'off',
         'no-unused-vars': ['error'],
     },
-    plugins: ['react'],
+    plugins: ["transform-class-properties","syntax-dynamic-import", "transform-object-rest-spread"],
 };
