+++ features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/store/pages/index.html
@@ -52,7 +52,7 @@
         </div>
     </div>
     <script type="text/javascript" src="/store/public/app/theme/defaultTheme.js"></script>
-    <script src="/store/public/app/dist/index.js"></script>
+    <script src="/store/public/app/dist/index.bundle.js"></script>
 
 </body>
