+++ features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/store/source/src/App.js
@@ -23,6 +23,7 @@ import {Apis, ApplicationCreate, Applications, Base, Login, Logout} from './app/
 import {PageNotFound} from './app/components/Base/Errors'
 import AuthManager from './app/data/AuthManager'
 import qs from 'qs'
+import Loadable from 'react-loadable';
 import ApplicationEdit from './app/components/Applications/Edit/ApplicationEdit';
 
 // import 'typeface-roboto'
@@ -35,166 +36,40 @@ import PrivacyPolicy from "./app/components/Policy/PrivacyPolicy";
 import CookiePolicy from "./app/components/Policy/CookiePolicy";
 import {addLocaleData, defineMessages, IntlProvider} from 'react-intl';
 import Configurations from "Config";
+import Progress from './app/components/Shared/Progress';
 
-const themes = [];
-
-themes.push(createMuiTheme(Configurations.themes.light));
-themes.push(createMuiTheme(Configurations.themes.dark));
-
-/**
- * Language.
- * @type {string}
- */
-const language = (navigator.languages && navigator.languages[0]) || navigator.language || navigator.userLanguage;
-
-/**
- * Language without region code.
- */
-const languageWithoutRegionCode = language.toLowerCase().split(/[_-]+/)[0];
-
-// import './materialize.css'
-
-/**
- * Render protected application paths
- */
-class Protected extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            showLeftMenu: false,
-            themeIndex: 0,
-            messages: {}
-        };
-        this.environments = [];
-        this.loadLocale = this.loadLocale.bind(this);
-        /* TODO: need to fix the header to avoid conflicting with messages ~tmkb*/
-
-    }
-
-    /**
-     * Load locale file.
-     *
-     * @param {string} locale Locale name
-     * @returns {Promise} Promise
-     */
-    loadLocale(locale = 'en') {
-        fetch(`${Utils.CONST.CONTEXT_PATH}/public/app/locales/${locale}.json`)
-            .then((resp) => resp.json())
-            .then((data) => {
-                // eslint-disable-next-line global-require, import/no-dynamic-require
-                addLocaleData(require(`react-intl/locale-data/${locale}`));
-                this.setState({messages: defineMessages(data)});
-            })
-    }
-
-    componentDidMount() {
-        ConfigManager.getConfigs().environments.then(response => {
-            this.environments = response.data.environments;
-            this.handleEnvironmentQueryParam();
-        }).catch(error => {
-            console.error('Error while receiving environment configurations : ', error);
-        });
-    }
-
-    /**
-     * Change the environment with "environment" query parameter
-     * @return {String} environment name in the query param
-     */
-    handleEnvironmentQueryParam() {
-        let queryString = this.props.location.search;
-        queryString = queryString.replace(/^\?/, '');
-        /* With QS version up we can directly use {ignoreQueryPrefix: true} option */
-        let queryParams = qs.parse(queryString);
-        const environmentName = queryParams.environment;
-
-        if (!environmentName || Utils.getEnvironment() === environmentName) {
-            // no environment query param or the same environment
-            return environmentName;
-        }
-
-        let environmentId = Utils.getEnvironmentID(this.environments, environmentName);
-        if (environmentId === -1) {
-            return environmentName;
-        }
-
-        let environment = this.environments[environmentId];
-        Utils.setEnvironment(environment);
-        return environmentName;
-    }
-    /**
-     * Change the theme index incrementally
-     */
-    componentWillMount() {
-        let storedThemeIndex = localStorage.getItem("themeIndex");
-        if (storedThemeIndex) {
-            this.setState({themeIndex: parseInt(storedThemeIndex)})
-        }
-        const locale = (languageWithoutRegionCode || language || 'en');
-        this.loadLocale(locale);
-    }
-    setTheme() {
-        this.setState({theme: themes[this.state.themeIndex % 3]});
-        this.state.themeIndex++;
-        localStorage.setItem("themeIndex", this.state.themeIndex);
-    }
-
-    render() {
-        const environmentName = this.handleEnvironmentQueryParam();
-        // Note: AuthManager.getUser() method is a passive check, which simply check the user availability in browser storage,
-        // Not actively check validity of access token from backend
-        if (AuthManager.getUser(environmentName)) {
-            return (
-                <IntlProvider locale={language} messages={this.state.messages}>
-                    <MuiThemeProvider theme={themes[this.state.themeIndex % 2]}>
-                        <Base setTheme={() => this.setTheme()}>
-                            <Switch>
-                                <Redirect exact from="/" to="/apis"/>
-                                <Route path={"/apis"} component={Apis}/>
-                                <Route path={"/applications"} component={Applications}/>
-                                <Route path={"/application/create"} component={ApplicationCreate}/>
-                                <Route path={"/application/edit/:application_id"} component={ApplicationEdit}/>
-                                <Route component={PageNotFound}/>
-                            </Switch>
-                        </Base>
-                    </MuiThemeProvider>
-                </IntlProvider>
-            );
-        } else {
-            return (
-                <IntlProvider locale={language} messages={this.state.messages}>
-                    <MuiThemeProvider theme={themes[this.state.themeIndex % 2]}>
-                        <Base setTheme={() => this.setTheme()}>
-                            <Route path={"/"} component={AnonymousView}/>
-                        </Base>
-                    </MuiThemeProvider>
-                </IntlProvider>
-            );
-        }
-
-        let params = qs.stringify({referrer: this.props.location.pathname});
-        return (
-            <Redirect to={{pathname: '/login', search: params}}/>
-        );
-    }
-}
+const LoadableProtectedApp = Loadable({
+    loader: () => import(// eslint-disable-line function-paren-newline
+        /* webpackChunkName: "ProtectedApp" */
+        /* webpackPrefetch: true */
+            './app/ProtectedApp'),
+    loading: Progress,
+});
 
 /**
  * Define base routes for the application
  */
 
-const Store = (props) => {
-    return (
-        <Router basename="/store">
-            <Switch>
-                <Route path={"/login"} render={() => <Login appName={"store"} appLabel={"STORE"}/>}/>
-                <Route path={"/logout"} component={Logout}/>
-                <Route path={"/sign-up"} component={SignUp}/>
-                <Route path={"/policy/privacy-policy"} component={PrivacyPolicy}/>
-                <Route path={"/policy/cookie-policy"} component={CookiePolicy}/>
-                <Route component={Protected}/>
-            </Switch>
-        </Router>
-    );
+class Store extends React.Component {
+    constructor(props) {
+        super(props);	 
+        LoadableProtectedApp.preload();
+    }
+    render() {
+        return (
+            <Router basename="/store">
+                <Switch>
+                    <Route path={"/login"} render={() => <Login appName={"store"} appLabel={"STORE"}/>}/>
+                    <Route path={"/logout"} component={Logout}/>
+                    <Route path={"/sign-up"} component={SignUp}/>
+                    <Route path={"/policy/privacy-policy"} component={PrivacyPolicy}/>
+                    <Route path={"/policy/cookie-policy"} component={CookiePolicy}/>
+                    <Route component={LoadableProtectedApp}/>
+                </Switch>
+            </Router>
+        );
+    }
+ 
 };
 
 export default Store;
